MDPs for Project certain outrrefs in the targetlist of subqueries
In a previous commit, we added code to fall back gracefully when a
subquery select list contains a single outer ref that is not part of an
expression, such as in

select * from foo where a is null or a = (select foo.b from bar)

This commit regenerates the MDPs for 6 test cases that were identified
having outer refs in their project lists. These test cases will go
through the fix in the translator and most of them will generate an
extra project in the DXL query section. Some of them now also have an
extra project in the plan section.

Co-authored-by: Hans Zeller <hzeller@pivotal.io>
Co-authored-by: Shreedhar Hardikar <shardikar@pivotal.io>