Disallow btree index scan on AO table
Method CXformUtils::FIndexApplicable() allowed a btree index to be
used for an IndexGet on an AO table, which is not supported.

Fixing the method to return false for this situation.

Now, transforms like CXformInnerJoin2DynamicIndexGetApply should no
longer generate an alternative for a btree index on an AO table.

This is a backport of master PR #11399.